refactor: extract shared build helpers from comp-detail.js
